rearchitect to put all the visualization code in one crate
