+++ GUIDE
@@ -6,12 +6,14 @@ enclone_args          argument processing
 enclone_client        highly experimental, minimal GUI client (not in use)
 enclone_com           inter-processs communication (experimenal)
 enclone_core          core definitions, etc.
+enclone_exec          just the actual main program
 enclone_help          help menus from the command line
-enclone_main          this includes the enclone main program
+enclone_main          this includes the enclone main program guts
 enclone_print         prints clonotypes in the sense of creating the "pictures" (much of core logic)
 enclone_proto         stuff for communicating with Loupe, and reused for other purposes
 enclone_tail          last part of the code
 enclone_tools         miscellaneous binaries
+enclone_visual        experimental GUI client
 
 credits               credits
 fonts                 DejaVuSansMono, in case not otherwise available
