+++ enclone_main/src/lib.rs
@@ -4,7 +4,6 @@ pub mod blacklist;
 pub mod determine_ref;
 pub mod disintegrate;
 pub mod doublets;
-pub mod enclone_client;
 pub mod fcell;
 pub mod filter_umi;
 pub mod flag_defective;
@@ -18,7 +17,3 @@ pub mod setup;
 pub mod split_orbits;
 pub mod subset;
 pub mod vars;
-
-pub mod proto {
-    tonic::include_proto!("enclone");
-}
